Revert "revert webkit patch to avoid overlapping on fixed layout book"
This reverts commit a05f4600901d0baef9e2bc1593bb7a974f1f8824.

+++ Source/WebCore/platform/graphics/Font.cpp
@@ -77,7 +77,7 @@ Font::Font()
 {
 }
 
-Font::Font(const FontDescription& fd, short letterSpacing, short wordSpacing)
+Font::Font(const FontDescription& fd, float letterSpacing, float wordSpacing)
     : m_fontDescription(fd)
     , m_letterSpacing(letterSpacing)
     , m_wordSpacing(wordSpacing)
